requirements.txt
IPython
google-genai
python-dotenv
